Checkpoint: Correcciones de APIs: 1. metrics.getAll: corregido input undefined→{} en Metrics.tsx y NutritionalStats.tsx 2. progress.summary: corregido error SQL 42883 añadiendo ::numeric cast en ROUND(AVG(...)) 3. admin.stats: ya corregido con totalAllergies y totalCategories 4. admin select accountType: corregido defaultValue→value (componente controlado) Todos los endpoints críticos verificados y funcionando correctamente.

+++ server/routers.ts
@@ -9282,10 +9282,10 @@ Devuelve SOLO JSON válido con esta estructura exacta:
       const userId = ctx.user.id;
       const [logsStats] = await drizzleDb.select({
         totalLogs: sql<number>`COUNT(*)`,
-        avgCalories: sql<number>`ROUND(AVG(${mealLogs.calories}), 0)`,
-        avgProteins: sql<number>`ROUND(AVG(${mealLogs.proteins}), 1)`,
-        avgCarbs: sql<number>`ROUND(AVG(${mealLogs.carbohydrates}), 1)`,
-        avgFats: sql<number>`ROUND(AVG(${mealLogs.fats}), 1)`,
+        avgCalories: sql<number>`ROUND(AVG(${mealLogs.calories})::numeric, 0)`,
+        avgProteins: sql<number>`ROUND(AVG(${mealLogs.proteins})::numeric, 1)`,
+        avgCarbs: sql<number>`ROUND(AVG(${mealLogs.carbohydrates})::numeric, 1)`,
+        avgFats: sql<number>`ROUND(AVG(${mealLogs.fats})::numeric, 1)`,
         daysWithLogs: sql<number>`COUNT(DISTINCT ${mealLogs.logDate})`,
       }).from(mealLogs).where(eq(mealLogs.userId, userId));
       const [firstMetric] = await drizzleDb.select({ weight: userMetrics.weight, date: userMetrics.date })
